fix: fix incorrect using of read method error in create, delete, update method

+++ amazon_rekognition/external_id_db.py
@@ -1,39 +1,33 @@
-from cgitb import enable
 from dataclasses import dataclass
-from distutils.log import error
-from boto3.dynamodb.conditions import Key
+from typing import Any, Optional
+
 import boto3
-from pprint import pprint
-from utils.measure_time import *
 
 
 class ExternalIdDb:
 
-    def __init__(self, dynamodb=None) -> None:
-        if not dynamodb:
-            dynamodb = boto3.resource('dynamodb', region_name='ap-northeast-2')
-        self.db = dynamodb
+    def __init__(self) -> None:
+        self.db = boto3.resource('dynamodb', region_name='ap-northeast-2')
 
     def create_table(self):
         table = self.db.create_table(
-        TableName='Faces',
-        KeySchema=[
-            {
-                'AttributeName': 'UserId',
-                'KeyType': 'HASH'  # Partition key
-            },
-        ],
-        AttributeDefinitions=[
-            {
-                'AttributeName': 'UserId',  #External Id 
-                'AttributeType': 'S',
-            },
-        ],
-        ProvisionedThroughput={
-            'ReadCapacityUnits': 10,
-            'WriteCapacityUnits': 10
-        }
-    )
+            TableName='Faces',
+            KeySchema=[
+                {
+                    'AttributeName': 'UserId',
+                    'KeyType': 'HASH'  # Partition key
+                },
+            ],
+            AttributeDefinitions=[
+                {
+                    'AttributeName': 'UserId',  #External Id 
+                    'AttributeType': 'S',
+                },
+            ],
+            ProvisionedThroughput={
+                'ReadCapacityUnits': 10,
+                'WriteCapacityUnits': 10
+            })
         return table
 
     def delete_table(self):
@@ -44,38 +38,34 @@ class ExternalIdDb:
         self,
         user_id: str,
         name: str,
-    )  -> bool:
-        read = self.read(user_id)
-        if not read :
-            return False  #ㅣ미 있는 것
+    ) -> bool:
+        is_exist: bool = False
+        try:
+            self.read(user_id)
+        except ExternalIdDbNotExistException:
+            is_exist = False
+
+        if not is_exist:
+            return False
 
         table = self.db.Table('Faces')
-        res = table.put_item(
-            Item = {
-                'UserId': user_id,
-                'UserName': name
-            }
-        )
+        res = table.put_item(Item={
+            'UserId': user_id,
+            'UserName': name,
+        })
         return True
-    
-    @measure_time
+
     def read(
         self,
         user_id: str,
     ):
         table = self.db.Table('Faces')
-        res = table.get_item(
-            Key={
-                'UserId': user_id
-            }
-        )
+        res = table.get_item(Key={'UserId': user_id})
 
         if not ('Item' in res):
-            #raise ValueError('없는 UserId 입니다.')
-            return False
+            raise ExternalIdDbNotExistException(user_id)
 
         return res['Item']['UserName']
-        
 
     def update(
         self,
@@ -84,32 +74,65 @@ class ExternalIdDb:
     ) -> bool:
 
         delete = self.delete(user_id)
-        if not delete :
-            return False         #업데이트 실행 X
 
-        res=self.create(user_id, new_name)
+        res = self.create(user_id, new_name)
         return True
-        
-    @measure_time
-    def delete(
-        self, 
-        user_id: str,
-    ) -> bool:  
 
-        read=self.read(user_id)
-        if not read : 
-            return False 
+    def delete(
+        self,
+        user_id: str,
+    ) -> bool:
+        is_exist: bool = False
+        try:
+            self.read(user_id)
+        except ExternalIdDbNotExistException:
+            is_exist = False
+
+        if not is_exist:
+            return False
 
         table = self.db.Table('Faces')
-        res = table.delete_item(
-            Key={
-                'UserId': user_id
-            },
-        )
+        res = table.delete_item(Key={'UserId': user_id}, )
         return True
+
+
+class ExternalIdDbException(Exception):
+
+    def __init__(
+        self,
+        message: str,
+        data: Optional[Any] = None,
+    ) -> None:
+        self.message = message
+        self.data = data
+
+    def __str__(self) -> str:
+        error_message: str = f'[ExternalIdDbError] {self.message}'
+
+        if not self.data:
+            error_message += f'\n{self.data}'
+
+        return error_message
+
+
+class ExternalIdDbAlreadExistException(ExternalIdDbException):
+
+    def __init__(self, user_id: str) -> None:
+        super().__init__(
+            'user_id is alread exist.',
+            {'user_id': user_id},
+        )
+
+
+class ExternalIdDbNotExistException(ExternalIdDbException):
+
+    def __init__(self, user_id: str):
+        super().__init__(
+            'user_id is not exist.',
+            {'user_id': user_id},
+        )
 
 
 if __name__ == '__main__':
-    enable_measure_time()
-    ex=ExternalIdDb()
+    ex = ExternalIdDb()
     ex.delete('11111')
